requirements.txt
# 기본 요구 사항
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.8.4
asyncio>=3.4.3

# 데이터베이스
sqlalchemy>=2.0.23
asyncpg>=0.28.0
psycopg2-binary>=2.9.9
langchain-postgres>=0.0.1

# LLM 및 자연어 처리
transformers>=4.35.2
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.22.0
bitsandbytes>=0.41.0  # 양자화 지원
google-generativeai>=0.3.1
tiktoken>=0.5.0  # 토큰 계산
langchain>=0.0.328

# API 연동
biopython>=1.81
kaggle>=1.5.16
pymed>=0.8.9

# 임베딩 및 벡터 검색
sentence-transformers>=2.2.2
openai>=1.3.3
faiss-cpu>=1.7.4  # 빠른 벡터 검색 (CUDA 가능한 환경에서는 faiss-gpu 사용 가능)
numpy>=1.24.2
scikit-learn>=1.2.2

# 성능 및 모니터링
psutil>=5.9.5  # 시스템 모니터링
prometheus-client>=0.17.0  # 메트릭 수집
memory-profiler>=0.61.0  # 메모리 사용 분석
py-spy>=0.3.14  # CPU 프로파일링

# 웹 및 UI
jinja2>=3.1.2
markdown>=3.4.3
pygments>=2.15.1

# 유틸리티
tqdm>=4.65.0  # 진행률 표시
loguru>=0.7.0  # 고급 로깅
tenacity>=8.2.2  # 재시도 로직
backoff>=2.2.1  # 지수 백오프

# 웹 스크래핑 및 크롤링
beautifulsoup4>=4.12.2
selenium>=4.15.2
scrapy>=2.11.0
lxml>=4.9.3
newspaper3k>=0.2.8
parsel>=1.8.1
googlesearch-python>=1.2.3
fake-useragent>=1.3.0
pyppeteer>=1.0.2

# 웹 검색
google-api-python-client>=2.108.0
bing-search>=1.1.2
duckduckgo-search>=3.9.9
serp-api>=0.1.0
searchengine-parser>=0.6.2

# 백터 DB
pgvector>=0.2.1
firebase-admin>=6.2.0
google-cloud-firestore>=2.12.0

# 의학 API 클라이언트
anthropic>=0.7.4
together>=0.2.8
huggingface-hub>=0.17.3
llama-cpp-python>=0.2.19